fixed other failures and moved out for pool
